fix: seven robustness fixes from the full-codebase review
A keyboard whose grab is refused no longer retries inside the device task's select arm. Looping there left events unread in the device buffer, and they were forwarded to the client after the grab landed -- input the local machine had already processed ungrabbed, delivered a second time. One attempt now returns to the select, where the existing retry_interval arm handles the failure and the task stays responsive.

'monux daemon switch' builds its request with serde_json rather than format!: a target containing a quote or a backslash produced a malformed request line. It failed closed and the socket is same-user, so this was a correctness wart rather than a hole -- and the last unescaped JSON construction left.

The device watcher's event channel is unbounded. Dropping a create or delete notice on a burst is unrecoverable: a missed Created leaves a keyboard unwatched for the life of the daemon, and these events are rare and tiny.

single_instance::live_holder opens the lock file with O_NOFOLLOW, matching the refusal acquire() already documents for world-writable /tmp. The probe decides the machine's role, which gates updates.

control.rs enforced 'line.len() > MAX_REQUEST_LINE + 1', letting a line through one byte past the documented cap.

maybe_elevate uses args_os(), like reexec_after_update: args() panics on non-UTF-8 argv.

transport::setsockopt returns () instead of a Result that could only ever be Ok, which removes a dead error branch and the close_fd() it guarded. The update lock passes truncate(false) explicitly: nothing reads the file, and truncating another updater's lock would be wrong.

+++ src/client.rs
@@ -782,7 +782,7 @@ impl Connection {
                     self.next_fetch_id = self.next_fetch_id.wrapping_add(1);
                     let msg = bulk::ClientBulk::ClipboardRequest(bulk::ClientClipboardRequest{
                         requested_type: &local_fetch_request.requested_type,
-                        max_size_bytes: self.max_clipboard_size_bytes as u64,
+                        max_size_bytes: self.max_clipboard_size_bytes,
                         request_id,
                     });
                     let serializedmsg = postcard::to_stdvec_cobs(&msg)
